+++ src/render/renderer_api.hpp
@@ -31,7 +31,7 @@ class RendererAPI
    Clear() = 0;
 
    virtual void
-   DrawIndexed(const std::shared_ptr< VertexArray >& vertexArray, uint32_t indexCount = 0) = 0;
+   DrawIndexed(const std::shared_ptr< VertexArray >& vertexArray, size_t indexCount = 0) = 0;
 
    virtual void
    DrawLines(uint32_t count) = 0;
